Mostly cleanup -removed console.log statements -changed file and component names
