--- src/client/routes/portraits.tsv
@@ -1,11 +1,0 @@
-filepath	name	role
-img/catnikolovski.jpeg	Catherine Nikolovski	Producer
-img/hobsonlane.jpeg	Hobson Lane	Lead Data Scientist
-img/kenwhaler.jpeg	Ken Whaler	Front End Team Lead
-img/danielmiller.jpeg	Daniel Miller	Back End Team Lead
-img/jamesofsink.jpeg	James Ofsink	Facilitator
-img/thundershiviah.jpeg	Thunder Shiviah	Data Scientist
-img/allengrimm.jpeg	Allen Grimm	Data Scientist
-img/colehoward.jpeg	Cole Howard	Data Scientist
-img/davidgarber.jpeg	David Garber	Front End Team
-img/lewisnotestine.jpeg	Lewis Notestine	Front End Team
